requirements/common.txt
PyMuPDF==1.25.5
streamlit==1.45.1
python-dotenv==0.21.0
PyYAML==6.0.1
openai==1.79.0